+++ models/nota_fiscal.py
@@ -11,8 +11,8 @@ from models.lote import Lote
 lotes_nota_fiscal = sa.Table(
     'lotes_nota_fiscal',
     ModelBase.metadata,
-    sa.Column('id_nota_fiscal', sa.Integer, sa.ForeignKey('notas_fiscais.id')),
-    sa.Column('id_lote', sa.Integer, sa.ForeignKey('lotes.id'))
+    sa.Column('id_nota_fiscal', sa.BigInteger, sa.ForeignKey('notas_fiscais.id')),
+    sa.Column('id_lote', sa.BigInteger, sa.ForeignKey('lotes.id'))
 )
 
 
@@ -28,7 +28,7 @@ class NotaFiscal(ModelBase):
     numero_serie: str = sa.Column(sa.String(45),unique=True, nullable=False)
     descricao: str = sa.Column(sa.String(200), nullable=False)
 
-    id_revendedor: Revendedor = sa.Column(sa.Integer, sa.ForeignKey('revendedores.id'))
+    id_revendedor: Revendedor = sa.Column(sa.BigInteger, sa.ForeignKey('revendedores.id'))
     revendedor: Revendedor = relationship('Revendedor', lazy='joined')
 
     #Uma nota fiscal pode ter vários lotes e um lote está ligado a uma nota fiscal
