+++ models/picole.py
@@ -13,22 +13,22 @@ from models.aditivo_nutritivo import AditivoNutritivo
 
 ingredientes_picole = sa.Table('ingredientes_picole',
     ModelBase.metadata,
-    sa.Column('id_ingrediente', sa.Integer, sa.ForeignKey('ingredientes.id')),
-    sa.Column('id_picole', sa.Integer, sa.ForeignKey('picoles.id'))
+    sa.Column('id_ingrediente', sa.BigInteger, sa.ForeignKey('ingredientes.id')),
+    sa.Column('id_picole', sa.BigInteger, sa.ForeignKey('picoles.id'))
     )
 
 conservantes_picole = sa.Table(
     'conservante_picole',
     ModelBase.metadata,
-    sa.Column('id_conservantes', sa.Integer,sa.ForeignKey('conservantes.id')),
-    sa.Column('id_picole', sa.Integer, sa.ForeignKey('picoles.id')) 
+    sa.Column('id_conservantes', sa.BigInteger,sa.ForeignKey('conservantes.id')),
+    sa.Column('id_picole', sa.BigInteger, sa.ForeignKey('picoles.id')) 
 )
 
 aditivos_nutritivos_picole = sa.Table(
     'aditivos_nutritivos_picole',
     ModelBase.metadata,
-    sa.Column('id_aditivo_nutritivo', sa.Integer, sa.ForeignKey('aditivos_nutritivos.id')),
-    sa.Column('id_picole', sa.Integer, sa.ForeignKey('picoles.id'))
+    sa.Column('id_aditivo_nutritivo', sa.BigInteger, sa.ForeignKey('aditivos_nutritivos.id')),
+    sa.Column('id_picole', sa.BigInteger, sa.ForeignKey('picoles.id'))
     )
 
 class Picole(ModelBase):
@@ -38,13 +38,13 @@ class Picole(ModelBase):
     preco: float = sa.Column(sa.DECIMAL(8,2), nullable=False)
     data_Criacao: datetime = sa.Column(sa.DateTime, default=datetime.now, index=True)
     
-    id_sabor: int = sa.Column(sa.Integer, sa.ForeignKey('sabores.id'))
+    id_sabor: int = sa.Column(sa.BigInteger, sa.ForeignKey('sabores.id'))
     sabor: Sabor = relationship('Sabor', lazy='joined') 
 
-    id_tipo_embalagem: int = sa.Column(sa.Integer, sa.ForeignKey('tipos_embalagem.id'))
+    id_tipo_embalagem: int = sa.Column(sa.BigInteger, sa.ForeignKey('tipos_embalagem.id'))
     tipo_embalagem: TipoEmbalagem = relationship('TipoEmbalagem', lazy='joined')
     
-    id_tipo_Picole: int = sa.Column(sa.Integer, sa.ForeignKey('picoles.id'))
+    id_tipo_Picole: int = sa.Column(sa.BigInteger, sa.ForeignKey('picoles.id'))
     tipo_picole: TipoPicole = relationship('TipoPicole', lazy='joined')
 
     ingredientes: List[Ingrediente] = relationship('Lote', secondary=ingredientes_picole, backref='ingredientes', lazy='dynamic')
